Separate process code into module
Create command module with console_script entry point
